backend/requirements.txt
beanie==1.26.0
fastapi==0.111.0
motor==3.4.0
passlib==1.7.4
pydantic==2.7.3
pydantic_settings==2.3.0
PyJWT==2.8.0
PyJWT==2.8.0
pytest==8.2.2
uvicorn==0.30.1
